// aws-connect-vm-portal/src/interface/agent.interface.ts
/******************************************************************************
 *  Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.                                           *                                                                                                                   *
 *  Licensed under the Apache License Version 2.0 (the 'License'). You may not
 *  use this file except in compliance with the License. A copy of the License
 *  is located at                                                            
 *                                                                                                                   
 *      http://www.apache.org/licenses/                                                                                   *                                                                                                                  
 *  or in the 'license' file accompanying this file. This file is distributed on
 *  an 'AS IS' BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, express or
 *  implied. See the License for the specific language governing permissions and
 *  limitations under the License.                                                                                
******************************************************************************/

export interface AgentInterface {
    userId: string;
    agent: any;
    connectUser: any;
}

export interface BasicAgentInterface {
    userId: string;
    username: string;
    extension: string;
    encryptVoicemail: boolean
    transcribeVoicemail: boolean
    deliveryOptions: any
}